feat(influxdb/flags): updated flags.yml with mergleFilterRule
Updated flags.yml to include a rule for mergeFilter and ran make flag to generate code for the flag. Added units to ensure the flag off and on behaviors. Also made sure that end_to_end_test.go ran the flux end to end test and passed the two merge filter tests.